yext/teamcity: move PropertyFromName to a Params method
Both Builds and Projects have a Params memeber, moving the PropertyFromName
function to be a member method of Params instead of Project means both
can use it with minimal refactoring.

The old version of the function will be kept to maintain compatability.
Properties is also added to the BuildType struct as it's returned for
build types but currently discarded.

R=atavakoli